requirements.txt
certifi==2019.9.11
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.15
django-appconf==1.0.3
django-bootstrap==0.2.4
django-imagekit==4.0.2
django-jet==1.0.7
gunicorn==19.9.0
idna==2.8
pilkit==2.0
Pillow==5.2.0
pkg-resources==0.0.0
psycopg2==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
pyuploadcare==2.5.0
requests==2.22.0
six==1.12.0
urllib3==1.25.4
whitenoise==4.1.4
